Improved quality of code and fixed some bugfixes.

+++ webservice-mvc/src/main/java/pl/konradboniecki/webservices/mvc/support/UserActivationController.java
@@ -1,7 +1,6 @@
 package pl.konradboniecki.webservices.mvc.support;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,13 +13,12 @@ import pl.konradboniecki.models.useractivationcode.UserActivationCodeRepository;
 
 import java.util.Optional;
 
-import static pl.konradboniecki.utils.enums.ErrorType.INVALID_ACTIVATION_LINK;
 import static pl.konradboniecki.templates.ViewTemplate.ERROR_PAGE;
+import static pl.konradboniecki.utils.enums.ErrorType.INVALID_ACTIVATION_LINK;
 
+@Log
 @Controller
 public class UserActivationController {
-
-    private static final Logger log = LoggerFactory.getLogger(UserActivationController.class);
 
     @Autowired
     private AccountRepository accountRepository;
@@ -32,19 +30,18 @@ public class UserActivationController {
                                      @PathVariable(name = "activationCode") String activationCodeFromUrl) {
 
         Optional<Account> acc = accountRepository.findById(id);
-        Optional<UserActivationCode> activationCode =
-                userActivationCodeRepository.findByAccountId(acc.get().getId());
         if (acc.isPresent()) {
+            Optional<UserActivationCode> activationCode =
+                    userActivationCodeRepository.findByAccountId(acc.get().getId());
             if (acc.get().isEnabled()) {
                 return new ModelAndView("redirect:/login");
             } else {
-                if (activationCode.isPresent()){
-                    if (activationCode.get().getActivationCode().equals(activationCodeFromUrl)){
+                if (activationCode.isPresent()
+                        && activationCode.get().getActivationCode().equals(activationCodeFromUrl)) {
                         accountRepository.setEnabled(id);
                         log.info("User with ID: " + acc.get().getId() + " has been activated");
                         userActivationCodeRepository.deleteById(activationCode.get().getId());
                         return new ModelAndView("redirect:/login");
-                    }
                 }
                 return new ModelAndView(ERROR_PAGE, "errorType", INVALID_ACTIVATION_LINK.getErrorTypeVarName());
             }
